corrections

+++ migrations/migrations/20190813164509-add-date-token.js
@@ -2,7 +2,7 @@
 module.exports = {
   up: (queryInterface, sequelize) =>
     queryInterface.addColumn('users', 'date_token', {
-      type: sequelize.INTEGER,
+      type: sequelize.BIGINT,
       allowNull: true
     }),
   down: queryInterface => queryInterface.removeColumn('users', 'date_token')
